Refactoring day
Change log:
1. Refactored alot of the code
2. Make the imports look cleaner
