+++ Space_Invaders/classes/Game/Screens/InstructionsMenuScreen.py
@@ -1,11 +1,5 @@
-try:
-    from .Screens import Screen
-    from .Enums import State
-    from .Colors import WHITE
-except ImportError:
-    from Screens import Screen
-    from Enums import State
-    from Colors import WHITE
+from . import Screen
+from .. import State, WHITE
 
 class InstructionsMenuScreen(Screen):
     def __init__(self, screen_width:int, screen_height:int, screen, debug:bool = False):
