+++ Space_Invaders/classes/Game/Screens/PauseScreen.py
@@ -1,15 +1,7 @@
 import pygame
 from pygame.locals import *
-
-try:
-    from .Screens import Screen
-    from .Enums import State, Direction
-    from .Colors import WHITE
-
-except ImportError:
-    from Screens import Screen
-    from Enums import State, Direction
-    from Colors import WHITE
+from . import Screen
+from .. import State, WHITE
 
 class PauseScreen(Screen):
     def __init__(self, screen_width:int, screen_height:int, screen, score:int, previous_state: State, debug:bool = False):
