+++ Space_Invaders/classes/Game/Screens/PlayScreen.py
@@ -1,24 +1,8 @@
 import pygame
 import random
 from pygame.locals import *
-try: 
-    from .Enums import State, Direction
-    from .Player import Player
-    from .EnemyGroup import EnemyShips
-    from .EnemyShip import EnemyShip
-    from .Colors import WHITE
-    from .Screens import Screen
-    from .Explosion import Explosion
-    from .Bullet import Bullet
-except ImportError:
-    from Enums import State, Direction
-    from Player import Player
-    from EnemyGroup import EnemyShips
-    from EnemyShip import EnemyShip
-    from Colors import WHITE
-    from Screens import Screen
-    from Explosion import Explosion
-    from Bullet import Bullet
+from . import Screen
+from .. import State, EnemyShips, Player, Direction, EnemyShip, WHITE, Bullet, Explosion
 
 class PlayScreen(Screen):
     def __init__(self, screen_width:int, screen_height:int, screen, sensitivity:int, max_fps:int, wave:int = 1, debug:bool = False):
@@ -60,7 +44,7 @@ class PlayScreen(Screen):
         self.down_bullets = pygame.sprite.Group()
 
         #Enemyships
-        self.enemies = EnemyShips()
+        self.enemies = EnemyShips(self.state)
 
         #Explosions
         self.explosions = pygame.sprite.Group()
