+++ Space_Invaders/classes/Game/Screens/__init__.py
@@ -1,0 +1,18 @@
+#Import all the screens
+from .Screens import Screen
+from .Popup import Popup
+from .PlayScreen import PlayScreen
+from .ClassicScreen import ClassicScreen
+from .GameoverScreen import GameoverScreen
+from .HighscoreScreen import HighscoreScreen
+from .InstructionsMenuScreen import InstructionsMenuScreen
+from .InstructionsScreen import InstructionScreen
+from .LocalPVPScreen import LocalPVPScreen
+from .MenuScreen import MenuScreen
+from .NewhighscoreScreen import NewhighscoreScreen
+from .PauseScreen import PauseScreen
+from .PlayModesScreen import PlayModeScreen
+from .PVPGameoverScreen import PVPGameoverScreen
+from .PVPInstructionsScreen import PVPInstructionsScreen
+from .PVPPauseScreen import PVPPauseScreen
+from .TwoPlayerScreen import TwoPlayerScreen
