+++ Space_Invaders/classes/Game/Sprites/EnemyGroup.py
@@ -1,13 +1,16 @@
 import pygame
+from .. import State
 
 class EnemyShips(pygame.sprite.Group):
-    def __init__(self):
+    def __init__(self, state: State):
         """The main class for the enemy ships group
             Arguments:
                 No arguments
         """
         #Initialise the superclass
         super().__init__()
+
+        self.state = state
 
     def update(self) -> None:
         """Update all the mobs within the group
@@ -18,7 +21,8 @@ class EnemyShips(pygame.sprite.Group):
         """
         if len(tuple(filter(lambda x: x.touch_edge(), self.sprites()))):
             for i in self.sprites():
-                i.move_down(i.get_height()//4)
+                if self.state != State.PVP:
+                    i.move_down(i.get_height()//4)
                 i.change_direction()
