+++ Space_Invaders/classes/Game/Sprites/ImageObject.py
