+++ Space_Invaders/classes/Game/Sprites/InputBox.py
@@ -1,9 +1,6 @@
 import pygame
 from pygame.locals import *
-try:
-    from .Colors import *
-except ImportError:
-    from Colors import *
+from .. import *
 
 class InputBox(object):
     def __init__(self, initial_x:int, initial_y:int, height:int, font, max_length:int = 10):
