+++ Space_Invaders/classes/Game/Sprites/MovingObject.py
