+++ Space_Invaders/classes/Game/Sprites/Object.py
@@ -1,8 +1,5 @@
 import pygame
 
-#Init pygame
-pygame.init()
-
 class Object(pygame.sprite.Sprite):
     def __init__(self, initial_x:int, initial_y:int, debug:bool = False):
         """Constructor for the object class in the game
