+++ Space_Invaders/classes/Game/Sprites/__init__.py
@@ -1,0 +1,12 @@
+from .Object import Object
+from .ImageObject import ImageObject
+from .MovingObject import MovingObject
+from .Background import Background
+from .Block import Block
+from .BlocksGroup import BlockGroup
+from .Bullet import Bullet
+from .EnemyGroup import EnemyShips
+from .EnemyShip import EnemyShip
+from .Explosion import Explosion
+from .Player import Player
+from .InputBox import InputBox
