+++ Space_Invaders/classes/Game/__init__.py
@@ -1,15 +1,19 @@
 #Import the Enums.py file
 from .Enums import *
 
+#Import the database
+from .database import *
+
+#Import the colors
+from .Colors import *
+
+#Import all the sprite classes
+from .Sprites import *
+
+#Import Screens
+from .Screens import *
+
 #Import the game.py
 from .game import *
 
-#Import playscreen
-from .PlayScreen import *
 
-#Import all the sprite classes
-from .Player import Player
-from .Bullet import Bullet
-from .Explosion import Explosion
-from .EnemyShip import EnemyShip
-from .Background import Background
